refactor(storage): host storage rules compilation with the resolution record
service.ts crossed the 600-line trigger in code-conventions.md section 2 with
the compilation step inlined in it. Compiling a source into a ruleset and its
resolution record is the concern rules-resolution.ts already owns, so
compileStorageRules moves there and both callers in service.ts read it from
one place.
